+++ cmd/rancid-cvs/main.go
@@ -11,7 +11,7 @@ import (
 	"gorancid/pkg/git"
 )
 
-const version = "0.4.2"
+const version = "0.4.3"
 
 func main() {
 	showVersion := flag.Bool("V", false, "print version")